Added buttons to load the Notes and QC files for each site
Added to the "By site" tab of the Data Summary page. 

Copied over all HMGZD_Notes and HMGZD_qc files to a new "www" folder on the repo. Simplest approach for the time being, but we'll have to update the files manually when new data is added. Manually adding files takes <1 minute, but perhaps not ideal. May be options to grab files thru googleDrive, but I'd expect that to be slow and much more complicated.